Add recipe template generator.
Must currently be run manually each time a recipe is added. This can be time consuming. A future version should templatize this page dynamically so that it can be updated and the database can be expanded without having to manually recreate the template each time. This should not adversely affect search performance because crawl bots receive the results of PHP outputs.